+++ packages/terra-paginator/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 2.76.0 - (August 25, 2021)
+
+* Added
+  * Added prop `pageLabel` to customize page label.
+
+* Changed
+  * Updated WDIO snapshots due to terra-functional-testing v2 upgrade.
+
 ## 2.75.0 - (July 6, 2021)
 
 * Changed